refactor: Properly distinguish between celestia-app and celestia-node
This changes the name of the existing `celestia` package to
`celestia-app` to better match the documentation and its `celestia-appd`
binary, and to reduce confusion around the `celestia` binary which is
actually included with the `celestia-node` package.

A wrapper `celestia` package has been included that wraps the
`celestia-app` and `celestia-node` packages under a single `celestia`
package using `symlinkJoin`. As a result, building `celestia` now
produces the same binaries that it used to (`celestia-appd`) while also
providing the actual `celestia` binary from the `celestia-node` package.